test: git-init: Test clones with submodules properly
The way how the submodules logic is tested is inadequate - the actual
logic isn't verified at all and only the log is inspected for a
particular line.
It wasn't until commit 6e7aca19db57ebfdaa2ac57841a1dd293c38ddc0 fixed
the submodule variable shadowing in the test that actually created a
test repo *with* submodules.

Fix that and verify that submodules were indeed initialized for the
artificial repository. Additionally, keep the log check, but move the
variable declaration & definition to the top of the test for ease of
reading.

Signed-off-by: Erik Skultety <eskultet@redhat.com>