git: Use '--single-branch' when in the shallow clone mode
A past patch made sure we reset the refspec before fetching the parent
repo data, but for git submodules we need a bit more.

This further limits the git objects that end up being fetched with
submodules, especially with shallow clones where we essentially don't
want anything apart the revision recorded in the parent repo.

Signed-off-by: Erik Skultety <eskultet@redhat.com>